use proper javax.measure.Quantity for hold volume
This is a bit overkill, given that hold volume is currently only used to
get the IATTC capacity class, but this is a bit of a test case for
potentially wider usage.

The Apache SIS library will also be useful for reading netCDF files, and
maybe some coordinate calculations, so it makes sense to bring it in.