fix FadMapTest
the original cause was that I was calling FadMap.deployFad directly in
the test, thus bypassing the FadManager. It would have been easy to work
around that, but it made me realize that the FadMap had no business
creating new FADs in the first place: it should be the FadManager's job,
so I refactored things accordingly.