enforce max trip duration in hours
see https://github.com/poseidon-fisheries/tuna/issues/24

I'm using the value for the longest trip we have in the dataset, so I
think it's unlikely to be be met anyway, but it does provide some safety
against aberrations. We'll see once we start calibrating what trip
lengths actually look like in practice.