remove "Random Plan" FAD fishing classes
I don't believe these are ever going to be used, and I don't like to let
cruft accumulate and get in the way. If we ever run scenarios that
require random behaviour for purse-seiners, it would probably work in a
different (simpler) way. Or we could always revert this commit if needed.