make sure purse seiner don't exceed their alloted hours
I had made FAD set duration stochastic in
7efc4265d8d1fa3afe1a7ded3682df7c5f560412
(ref: https://github.com/poseidon-fisheries/tuna/issues/7)

...but I had not taken that properly into account when checking if there
was enough time to make a set. The rng was called at the time of checking
and again at the time to make the set, so it was possible that the actual
time spent ended up longer than when first checked.

The duration is now stored at the time of action instantiation instead.