separate success probability of FAD sets and unassociated sets
since the unassociated sets are going to be sampled from a distribution
file that includes unsuccessful sets.

It was either that or maintaining a separate success probability for
both types of sets, but this was the easiest to implement and I'm in a
bit of a rush. That decision might be worth revisiting later.­..