remove jgit from the build
as it was somehow breaking the creating of fat jars
(and thus preventing the optimizer to run from the console)

I believe the issue is related to
https://stackoverflow.com/questions/51455197/gradle-fatjar-could-not-find-or-load-main-class

but adding exclude `'META-INF/*.RSA', 'META-INF/*.SF', 'META-INF/*.DSA'`
didn't fix it. I think that syntax is just base Gradle, however, and doing
the same with the Shadow plugin might require a slightly different syntax.

It would be worth fiddling with this a bit more, as having the commit
hash tied to optimization results is super useful, but I'm willing to
give this up right now just to get things running again.