remove Route.equals method
I realized having an equals method for a mutable object might lead to
tricky bugs, so I removed it (for now). This makes the test code uglier
but I can live with that for now.

I'm leaving the equalsverifier dependency in, though, as we should
definitely test all equals methods in our codebase...