convert CurrentVectors to use Int2D instead of SeaTile
the idea being to decouple it from the actual map so we can share it
between parallel runs.

The CurrentVectorsFactory is now a singleton that maintains a cache of
CurrentVectors objects for maps that share the same MapExtent (a new
value class storing basic map properties) and current files. I'd be
surprised if we ever used different current maps within the same set of
runs, but the current structure supports it.