I have been keeping these two tests failing for now a month because I was quite unsure whether I should be solving them. But I think at this point I need to make a decision.
(1) The counter shouldn't count stuff without adding a column first, but it is used so often that checking is more expensive than I want to. The check happens when you ask the count back and I think it's fair to assume that this is when the exception will be thrown
(2) The literature on how to apply ITarget is a bit weird because different sources apply the multiplier the initial time step or only from t=2; the test I wrote assumed the multiplier applied at t=1 but the class had since assumed the change occurs at t=2. Still unsure about this one but for now I will make the multiplier apply immediately