fix "tooFewBins" test
This test was broken in 3bcd9b3170a1894117058372d67c87fe65a4fede.

The exception was changed to a logged warning, so I had to capture the
logger output to test it. I also had to change the number of bins to get
the message to occur. Not entirely sure that was the right approach,
though. @CarrKnight might want to have a look.