help the StrategyFactoryInspector handle edge cases
The code is still a bit convoluted, but it covers all the cases we
currently have in POSEIDON without resorting to string parsing
shenanigans.

I feel like I should have commented the code, but I can't seem to find
a middle between "just reformulate what the code is doing" and "write
a ten page essay on Java reflection involving generics".