add exogenous mortality and perform recruitment on aggregated biomass
A big commit with some code maintenance, but the two main changes are
the ones mentioned above.

We now apply recruitment to the aggregated (ocean wide) abundance, since
we realized that the R0 would otherwise need to be scaled.

And we add exogenous mortality because it was long overdue and because
@brianreedpowers needs it to test the combined biological processes.

Note that the natural and exogenous mortality are handled by the
same AbundanceMortalityProcess class. That means we can probably remove
the mortality values from TunaMeristics, but I haven't done that yet.