allow schools sets to grab fish from FADs if not enough in the cell
Important bits of logic are in:

- SchoolSetOpportunityGenerator, where the biology considered in order
to generate a set or not is now an aggregate of cell and FAD biologies.

- SchoolSetAction, where a successful set will first grab fish from the
ocean and then from the FADs (in random order) until the correct catch
is generated.

- AbundanceCatchMaker, which implements the logic of "grabbing as much
as we can" and return "caught" and "uncaught" fish as a pair. There is
a biomass version of that class too.