optimize abundance biology aggregation
The need for this comes from the fact that we now generate school sets
based on the aggregation of cell biology + FAD biologies, which means
that we're calling this way more often, so we need to make it faster.

Also, when aggregating biologies, we were loosing the biomass info,
which is costly to recalculate. We now take advantage of the precomputed
biomasses from the biologies that we're aggregating if they're available.