use weight groups in abundance reallocator factory too
Now that we have a proper notion of weight groups, we can avoid
duplicating the idea with the "firstLargeBin" field that we had
in the abundance reallocator.

There is a bit of a problem still because the weight we are currently
using for SKJ don't go all the way to 15kg, which was previously our
cutoff for "large" fish used to choose which distribution map to use.

I temporarily said that the large weight group for SKJ starts at 11.5019
but we'll need to sort that out.