make sure SpecificProtectedAreas are initialized correctly
The main change is that those regulations don't rely on the SeaTile
having a reference to the MPA's geometry object anymore. Every protected
area has an `inArea` array, indexed by grid coordinates, and that's all
that matters. That makes it easy to share areas between simulations and
saves us from having to call `recomputeTilesMPA` on the map when they're
initialized.

The main downside is that we don't see areas as grey zones in the GUI
anymore, but that's fixable, albeit low-priority.

I also added a pair of test that checks, for a set of points including
one in each EEZ, that regulations are applied correctly. That made me
realize that GeoMason was reading only the first polygon in multipolygon
shapes. I fixed that upstream (in
https://github.com/poseidon-fisheries/tuna/commit/e9fdbe26ced2f28d72b1615c882bcfa4200c41d7
) by taking only the largest polygon in each EEZ.