remove maximumAttractionDays
that was the cause of the bug mentioned in
https://github.com/poseidon-fisheries/tuna-issues/issues/141#issuecomment-1511944183

In short, I had set maximumDaysAttraction to Integer.MAX_VALUE in order
to keep attracting indefinitely. The problem was caused by the fact that
the CatchabilitySelectivityFishAttractor added this number to
daysInWaterBeforeAttraction, hence causing an integer overflow and
always failing the condition for attraction.

Since I don't believe that FADs should ever stop attracting anyway, I
removed the condition and the parameter altogether. We can always put
it back if needed but I'd need a lot of convincing...