make `BetterThanAverageEroteticDestinationFactoryTest` deterministic
We have a bunch of tests that rely on randomness and, consequently,
fail once in a while. That's pretty since we're now using protected
branches on GitHub and can only merge if all tests pass.

To get around that, I think it's all right to used fixed seeds for this
kind of tests. I'll fix them as I come across them...