eliminate compiler warnings in main POSEIDON code
cleaning up some code but mostly using SuppressWarnings annotations,
the goal being to allow compiling with `-Werror` and enforcing cleaner
new code.

This is a massive commit and it would be surprising if it didn't have
some undesirable consequences here and there, but that's something I'm
willing to live with.