remove constraint on minimum FAD value in VPS module
we _want_ a minimum FAD value, because we are trying to avoid generating
too many sets with a value of zero. The constraint of not having a
minimum FAD value in this particular module was based on the idea that
zeros should be accounted for in the average "value per set" calculation
but I see that as debatable: if you're trying to pick a region, why
would two FADs worth 100K$ be better than two FADs worth 100K$ _and_
a FAD worth zero?

More pragmatically, there is no way to separate the FADs considered for
the choice of region and the FADs considered for making a set, so if we
want to avoid picking zero-valued FADs, we need to exclude them from the
region choice calculation as well. (Or to overhaul the whole thing...)