push getCarryingCapacities down to PerSpeciesCarryingCapacity interface
and add argument checks to a few places that actually use this method.

It would be better to either have those classes rely on the isFull
method or statically ensure that they only deal with
PerSpeciesCarryingCapacity but none of these are actually used at the
moment so it's not worth the effort.