refactor environmental map factories
There was some weird duplication with two different classes called
EnvironmentMapFactory doing almost (but not quite) the same thing.

This is a pretty hefty refactoring, as I moved all of that logic to
a new `geography` sub-project with a bunch of new factories.

There is no time for that now, but I believe a lot of the new code could
be reused for, say, the species distributions maps.