be smarter about how we compute parameter ranges for local calibration
Before, we were using calibrated value ±x%, we meant that values near to
hard maximum got a much bigger range than values near the minimum.

Now, we take x% of the range between hard min and max, and center that
on the calibrated value, so the effect is the same no matter where we
are in the global range.

Also the new range is capped at the hard minimum or maximum, but we
don't truncate it, so the calibrated value won't be centered when it's
very close to the hard bounds. For example, if the hard bounds are 0-1
and we want a range of 10% around 0.05, the soft bounds will be 0-0.1.