use snapshot version of gdx-ai
because the latest official release (1.8.2) uses an outdated version
of libgdx that had a bug in the Bresenham2 pathfinder. The libgdx
dependency has been upgraded in a more recent commit, but gdx-ai 1.8.3
has never been released.

It would be good to move away from the currently unmaintained gdx-ai
eventually, but while there are plenty of pathfinding implementations
out there, there are no actively maintained behaviour tree libraries
for Java.